+++ azure-kusto-ingest/setup.py
@@ -46,10 +46,8 @@ setup(
         "azure-storage-blob>=1.1.0",
         "azure-storage-common>=1.1.0",
         "azure-storage-queue>=1.1.0",
-        "six>=1.10.0",      
+        "six>=1.10.0",
     ],
-        extras_require={
-        "pandas": ["pandas>=0.15.0"]
-    },
+    extras_require={"pandas": ["pandas>=0.15.0"]},
     cmdclass=cmdclass,
 )
